fix(learning-units-graph): rename folder
